Reorganize backend folder structure
