+++ package.json
@@ -26,6 +26,7 @@
   "homepage": "https://github.com/Star-Nosed/Moles#readme",
   "dependencies": {
     "axios": "^0.25.0",
+    "bcrypt": "^5.0.1",
     "bootstrap": "^5.1.3",
     "copy-to-clipboard": "^3.3.1",
     "css-loader": "^6.5.1",
@@ -35,6 +36,7 @@
     "install": "^0.13.0",
     "nodemon": "^2.0.15",
     "npm": "^8.4.0",
+    "pg": "^8.7.1",
     "react": "^17.0.2",
     "react-copy-to-clipboard": "^5.0.4",
     "react-dom": "^17.0.2",
