+++ server/controllers/dbController.js
@@ -1,0 +1,1 @@
+const db = require('../models/dbModel');
