+++ server/secrets/secrets.js
@@ -1,0 +1,5 @@
+const YELP_KEY = 'Bearer kDH3a4z30wEffzgX0iyn8OTTyqvuVU4zUw9vPFzfFi9p8pFJxtEeyWyoDH1hYi2jpNNUnYmhrtu1OrwI1Q_mOMZwQbTY95bKtm8IN-xynKO82AHLjd_CS2fjfdr5YXYx';
+
+const DB_URL = 'postgresql://noodle:BGothCYGUeNmAZ4f@db-postgresql-nyc1-82145-do-user-10576934-0.b.db.ondigitalocean.com:25061/noodledb-pool1';
+
+module.exports = { YELP_KEY, DB_URL };
