+++ server/secrets/secrets_blank.js
@@ -1,0 +1,4 @@
+const YELP_KEY = 'yelp API key goes here';
+
+const DB_URL = 'db connection string goes here';
+
